requirements.txt
# AlphaAgent - Deep Reinforcement Learning Trading Agent
# Complete system with all advanced features

# Core ML/DRL
torch>=2.0.0
stable-baselines3>=2.0.0
gymnasium>=0.29.0
numpy>=1.24.0
pandas>=2.0.0

# Financial Data
yfinance>=0.2.28
ta>=0.11.0  # Technical analysis library
pandas-ta>=0.3.14b  # Additional technical indicators

# NLP & Sentiment Analysis
transformers>=4.30.0
openai>=1.0.0
sentencepiece>=0.1.99

# Image transformation for GAF
pyts>=0.12.0
scikit-image>=0.21.0

# Machine Learning & Analysis
scikit-learn>=1.3.0
scipy>=1.11.0

# Explainability
shap>=0.42.0  # SHAP feature attribution
lime>=0.2.0.1  # LIME local explanations

# Visualization & Analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Dashboard
dash>=2.14.0
dash-bootstrap-components>=1.5.0

# Production & Monitoring
tensorboard>=2.14.0
rich>=13.5.0  # Beautiful terminal output
tqdm>=4.65.0

# Utils
python-dotenv>=1.0.0
requests>=2.31.0  # For API calls
